Split up main file
