+++ lib/device_wizard/browser_details.rb
@@ -1,0 +1,11 @@
+module DeviceWizard
+  class BrowserDetails
+    attr_accessor :name
+    attr_accessor :version
+
+    def initialize
+      @name = UNKNOWN
+      @version = UNKNOWN
+    end
+  end
+end
