+++ lib/device_wizard/device_type.rb
@@ -1,0 +1,9 @@
+module DeviceWizard
+  class DeviceType
+    DESKTOP = 'desktop'
+    TABLET  = 'tablet'
+    MOBILE  = 'mobile'
+    CRAWLER = 'crawler'
+    UNKNOWN = 'unknown'
+  end
+end
